recommendation/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
python-dotenv==1.0.1 